collect workout from google fit

+++ android/src/main/kotlin/com/metaflow/flutterhealthfit/WorkoutsReader.kt
@@ -88,7 +88,7 @@ class WorkoutsReader {
             "end" to end,
         )
         if (energy != null) {
-            map["energy"] = energy.toInt()
+            map["energy"] = energy
         }
         if (source != null) {
             map["source"] = source
